Add feature to initialize map on ShowOneRun
Incomplete feature.

Trying to rebuild a map2 object within ui.js, so that when I get a
successful SHOW, I can rebuild a map to show that route.

As of right now, its initializing with data.run.route[0] as the starting
point, but google doesn't like it as a string.  Need to figure out a way
to turn it into a number.